# packages/ts/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "DescriptionIn": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "maxLength": 200,
            "title": "Description",
            "type": "string"
          },
          "version": {
            "minimum": 0.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "description"
        ],
        "title": "DescriptionIn",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "GroupAdminOut": {
        "additionalProperties": false,
        "properties": {
          "agents": {
            "items": {
              "$ref": "#/components/schemas/GroupAgentOut"
            },
            "title": "Agents",
            "type": "array"
          },
          "area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "fixed_type": {
            "title": "Fixed Type",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "system": {
            "title": "System",
            "type": "boolean"
          },
          "type": {
            "enum": [
              "central",
              "area",
              "general"
            ],
            "title": "Type",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "type",
          "area",
          "description",
          "version",
          "system",
          "fixed_type",
          "agents"
        ],
        "title": "GroupAdminOut",
        "type": "object"
      },
      "GroupAgentOut": {
        "additionalProperties": false,
        "properties": {
          "account_data": {
            "title": "Account Data",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "account_data"
        ],
        "title": "GroupAgentOut",
        "type": "object"
      },
      "GroupChangeOut": {
        "additionalProperties": false,
        "properties": {
          "after": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroupTargetOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "agents": {
            "title": "Agents",
            "type": "integer"
          },
          "before": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroupShapeOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "change_id": {
            "title": "Change Id",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By"
          },
          "decided_by_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By Email"
          },
          "expires_at": {
            "title": "Expires At",
            "type": "string"
          },
          "group_id": {
            "title": "Group Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "create",
              "update",
              "delete"
            ],
            "title": "Kind",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "proposed_by": {
            "title": "Proposed By",
            "type": "string"
          },
          "proposed_by_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed By Email"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "approved",
              "rejected",
              "withdrawn",
              "expired"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "change_id",
          "kind",
          "group_id",
          "status",
          "before",
          "after",
          "agents",
          "proposed_by",
          "proposed_by_email",
          "reason",
          "created_at",
          "expires_at",
          "decided_by",
          "decided_by_email",
          "decided_at",
          "note"
        ],
        "title": "GroupChangeOut",
        "type": "object"
      },
      "GroupList": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GroupOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "GroupList",
        "type": "object"
      },
      "GroupOut": {
        "additionalProperties": false,
        "properties": {
          "area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "area",
          "description"
        ],
        "title": "GroupOut",
        "type": "object"
      },
      "GroupShapeOut": {
        "additionalProperties": false,
        "properties": {
          "area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area"
          },
          "type": {
            "enum": [
              "central",
              "area",
              "general"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "area"
        ],
        "title": "GroupShapeOut",
        "type": "object"
      },
      "GroupTargetOut": {
        "additionalProperties": false,
        "properties": {
          "area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "type": {
            "enum": [
              "central",
              "area",
              "general"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "area",
          "description"
        ],
        "title": "GroupTargetOut",
        "type": "object"
      },
      "GroupsAdminOut": {
        "additionalProperties": false,
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/GroupChangeOut"
            },
            "title": "Changes",
            "type": "array"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/GroupAdminOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "changes"
        ],
        "title": "GroupsAdminOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "mango-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/groups": {
      "get": {
        "operationId": "get_admin_groups",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupsAdminOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Get Admin Groups"
      }
    },
...
    "/admin/groups/{group_id}/description": {
      "put": {
        "operationId": "put_group_description",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9-]{1,63}$",
              "title": "Group Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DescriptionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupsAdminOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Put Group Description"
      }
    },
...
    "/groups": {
      "get": {
        "operationId": "list_groups",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupList"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "List Groups"
      }
    },
...
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}